requirements.txt
asgiref==3.2.7
Django==3.0.6
djangorestframework==3.10.3
pytz==2020.1
sqlparse==0.3.1
